sdk now generates JSON Schema for later uses
