+++ plugin-sdk-go.go
@@ -39,7 +39,7 @@ func NewGenerator(specPath, packageRoot string, force bool) (*Generator, error) 
 		Triggers:     make(map[string]PluginHandlerData),
 		Connection:   make(map[string]ParamData),
 	}
-	if err := yaml.Unmarshal(data, s); err != nil {
+	if err = yaml.Unmarshal(data, s); err != nil {
 		return nil, err
 	}
 	// Fill in some basic stuff that isn't readily available from the parse, but helps the generation
@@ -63,6 +63,7 @@ type PluginSpec struct {
 	Help                          string                         `yaml:"help"`
 	Connection                    ParamDataCollection            `yaml:"connection"`
 	ConnectionRequiresCustomTypes bool                           `yaml:"-"`
+	ConnectionSchema              JSONSchema                     `yaml:"-"`
 	RawTypes                      map[string]ParamDataCollection `yaml:"types"`
 	Types                         map[string]TypeData            `yaml:"-"`
 	Triggers                      PluginHandlerCollection        `yaml:"triggers"`
@@ -76,130 +77,6 @@ type PluginSpec struct {
 	TypeMapper        *TypeMapper `yaml:"-"`
 	SpecLocation      string      `yaml:"-"`
 	ConnectionDataKey string      `yaml:"-"`
-}
-
-// ParamData are the details that make up each input/output/trigger param
-// Not all data is always filled in, it's context sensitive, which is unfortunate but
-// I'm willing to accept it for this since everything else is greatly simplified
-type ParamData struct {
-	RawName     string        `yaml:"name"`
-	Name        string        `yaml:"-"` // This is the joined and camelled name for the param
-	RawType     string        `yaml:"-"` // This is the original type as it was in the yaml
-	Type        string        `yaml:"type"`
-	Required    bool          `yaml:"required"`
-	Description string        `yaml:"description"`
-	Enum        []interface{} `yaml:"enum"`
-	Default     interface{}   `yaml:"default"`
-	Embed       bool          `yaml:"embed"`
-	Nullable    bool          `yaml:"nullable"`
-	Order       int
-	Schema      JSONSchema
-
-	// Things that are used for background help
-	EnumLiteral []EnumData `yaml:"-"`
-}
-
-// ParamDataCollection defines a type we can control marshal behavior with
-type ParamDataCollection map[string]ParamData
-
-// UnmarshalYAML lets us do some magic to set an order since go maps lack it
-func (pdc *ParamDataCollection) UnmarshalYAML(unmarshal func(v interface{}) error) error {
-	if pdc == nil {
-		pdc = &ParamDataCollection{}
-	}
-	m := make(map[string]ParamData)
-	if err := unmarshal(&m); err != nil {
-		return fmt.Errorf("Unable to unmarshal yaml into ParamDataCollection: %s", err)
-	}
-	// Unmarshal again into an ordered list (in order to set Order property)
-	orderedParameters := make(yaml.MapSlice, 0)
-	if err := unmarshal(&orderedParameters); err != nil {
-		return fmt.Errorf("Unable to unmarshal yaml into ordered parameters: %s", err)
-	}
-	for i, p := range orderedParameters {
-		key, ok := p.Key.(string)
-		if !ok {
-			return fmt.Errorf("Unable to unmarshal yaml parameters: Parameter name was not a string")
-		}
-		val, ok := m[key]
-		if !ok {
-			return fmt.Errorf("Could not unmarshal yaml parameter: %s key was somehow not in ParamDataCollection", key)
-		}
-		val.RawType = val.Type
-		val.RawName = key
-		val.Name = UpperCamelCase(key)
-		val.Order = i + 1 // Set the order
-		m[key] = val
-	}
-	*pdc = m
-	return nil
-}
-
-// EnumData is used to parse and write out enums
-type EnumData struct {
-	Name         string `yaml:"-"`
-	LiteralValue string `yaml:"-"`
-}
-
-// TypesInternalType is a special hack for the types package. Because in all other places we need to reference X via types.X
-// we prefix it ahead of time. But types that use or refer to other types in the types package don't need it.
-// This method, used only from the code generators for types, is to make sure types don't use the pacakge name internally in the package
-func (p ParamData) TypesInternalType() string {
-	if i := strings.Index(p.Type, "types."); i > -1 {
-		return strings.Replace(p.Type, "types.", "", -1)
-	}
-	return p.Type
-}
-
-// PluginHandlerData defines the actions or triggers
-type PluginHandlerData struct {
-	RawName             string `yaml:"name"`
-	Name                string `yaml:"-"` // This is the joined and camelled name for the action
-	Title               string `yaml:"title"`
-	Description         string `yaml:"description"`
-	Input               ParamDataCollection
-	Output              ParamDataCollection
-	RequiresCustomTypes bool // Used to assist in generating correct imports
-	Order               int  // Used to preserve the exact ordering in the spec
-	Schema              JSONSchema
-	// Things that are only used to make parsing templates simpler
-	PackageRoot string `yaml:"-"`
-	HasInterval bool   `yaml:"-"`
-}
-
-// PluginHandlerCollection defines a type we can control marshal behavior with
-type PluginHandlerCollection map[string]PluginHandlerData
-
-// UnmarshalYAML lets us do some magic to set an order since go maps lack it
-func (phc *PluginHandlerCollection) UnmarshalYAML(unmarshal func(v interface{}) error) error {
-	if phc == nil {
-		phc = &PluginHandlerCollection{}
-	}
-	m := make(map[string]PluginHandlerData)
-	if err := unmarshal(&m); err != nil {
-		return fmt.Errorf("Unable to unmarshal yaml into PluginHandlerCollection: %s", err)
-	}
-	// Unmarshal again into an ordered list (in order to set Order property)
-	op := make(yaml.MapSlice, 0)
-	if err := unmarshal(&op); err != nil {
-		return fmt.Errorf("Unable to unmarshal yaml into ordered parameters: %s", err)
-	}
-	for i, p := range op {
-		key, ok := p.Key.(string)
-		if !ok {
-			return fmt.Errorf("Unable to unmarshal yaml parameters: Parameter name was not a string")
-		}
-		val, ok := m[key]
-		if !ok {
-			return fmt.Errorf("Could not unmarshal yaml parameter: %s key was somehow not in PluginHandlerCollection", key)
-		}
-		val.RawName = key
-		val.Name = UpperCamelCase(key)
-		val.Order = i + 1 // Set the order
-		m[key] = val
-	}
-	*phc = m
-	return nil
 }
 
 // TypeData defines the custom types. Much of the data is pulled via a parent-key, so we don't parse much from yaml at all.
@@ -217,126 +94,6 @@ type HTTP struct {
 	Port         int `yaml:"port"`
 	ReadTimeout  int `yaml:"read_timeout"`
 	WriteTimeout int `yaml:"write_timeout"`
-}
-
-// PostProcessSpec does some minor post-processing on the spec object to fill a few things in that make
-// template generation easier
-func PostProcessSpec(s *PluginSpec) error {
-	// Create a new TypeManager and feed it some metadata about the spec, for it's own benefits
-	t := NewTypeMapper(s)
-	s.TypeMapper = t
-
-	// Handle any custom types
-	// We'll both populate Types AND update RawTypes so the original source is correct w/r/t the downstream source
-	s.Types = make(map[string]TypeData)
-	for name, data := range s.RawTypes {
-		td, err := t.RawTypeToType(name, data)
-		if err != nil {
-			return err
-		}
-		s.RawTypes[name] = data
-		s.Types[name] = *td
-	}
-	// now fill in all the schemas
-	if err := t.PopulateSchemas(s.Types); err != nil {
-		return err
-	}
-
-	// fill in the connection names
-	// Do this one out long form since we need the special case of building the data key
-	for name, param := range s.Connection {
-		param.RawName = name
-		param.Name = UpperCamelCase(name)
-		specType := param.Type
-		param.Type = t.SpecTypeToGoType(param.Type)
-		param.RawType = specType
-		// Make sure to import the plugins own types package when generating the connection
-		if !s.ConnectionRequiresCustomTypes && strings.Contains(param.Type, "types.") {
-			s.ConnectionRequiresCustomTypes = true
-		}
-		s.Connection[name] = param
-		// This is all stupid hacky but we need some way to hash the params and have a consistent key
-		// for the connection in the hash for looking up via the data incoming with the message
-		if param.Type == "string" && specType != "python" && specType != "password" {
-			if s.ConnectionDataKey != "" {
-				s.ConnectionDataKey += " + "
-			}
-			s.ConnectionDataKey += "c." + param.Name
-		}
-	}
-	if s.ConnectionDataKey == "" {
-		// Default it to the literal value of an empty string for now
-		// This could be because there were no string params - an issue to solve
-		// Or because it doesn't use a connection, which is totally fine
-		// TODO if there is no connection to generate, skip the whole connection pkg?
-		s.ConnectionDataKey = `""`
-	}
-	// fill in the action and trigger data
-	for name, action := range s.Actions {
-		if action.RawName != "" && action.Title == "" {
-			// Very old plugins used name, not title
-			// We use RawName for different purposes in the gosdk
-			// swap the values around if no title, but name was set
-			action.Title = action.RawName
-		}
-		action.RawName = name // not set in the yaml this way, but set for the benefit of the template
-		action.Name = UpperCamelCase(name)
-		action.PackageRoot = s.PackageRoot
-		// We need to do the same thing for the params too
-		customTypesInput, err := t.updateParams(action.Input)
-		if err != nil {
-			return err
-		}
-		customTypesOutput, err := t.updateParams(action.Output)
-		if err != nil {
-			return err
-		}
-		if customTypesInput || customTypesOutput {
-			action.RequiresCustomTypes = true
-		}
-		s.Actions[name] = action
-	}
-
-	for name, trigger := range s.Triggers {
-		if trigger.RawName != "" && trigger.Title == "" {
-			// Very old plugins used name, not title
-			// We use RawName for different purposes in the gosdk
-			// swap the values around if no title, but name was set
-			trigger.Title = trigger.RawName
-		}
-		trigger.RawName = name // not set in the yaml this way, but set for the benefit of the template
-		trigger.Name = UpperCamelCase(name)
-		trigger.PackageRoot = s.PackageRoot
-		// We need to do the same thing for the params too
-		customTypesInput, err := t.updateParams(trigger.Input)
-		if err != nil {
-			return err
-		}
-		customTypesOutput, err := t.updateParams(trigger.Output)
-		if err != nil {
-			return err
-		}
-		if customTypesInput || customTypesOutput {
-			trigger.RequiresCustomTypes = true
-		}
-		if _, ok := trigger.Input["interval"]; ok {
-			trigger.HasInterval = true
-		}
-		s.Triggers[name] = trigger
-	}
-
-	if s.HTTP.Port == 0 {
-		s.HTTP.Port = 10001
-	}
-
-	if s.HTTP.ReadTimeout == 0 {
-		s.HTTP.ReadTimeout = 60 * 10 // Default read timeout is 10 mins
-	}
-
-	if s.HTTP.WriteTimeout == 0 {
-		s.HTTP.WriteTimeout = 60 * 10 // default write timeout is 10 mins
-	}
-	return nil
 }
 
 // GeneratePlugin emits the new plugin based on the spec to disk
@@ -594,7 +351,6 @@ func (g *Generator) runGoImports() error {
 	if err != nil {
 		return err
 	}
-
 	for _, p := range fileList {
 		cmd := exec.Command("goimports", "-w", "-srcdir", g.spec.PackageRoot, p)
 		if b, err := cmd.CombinedOutput(); err != nil {
@@ -610,19 +366,18 @@ func (g *Generator) vendorPluginDeps() error {
 	rootPath := path.Join(os.Getenv("GOPATH"), "/src/", g.spec.PackageRoot)
 	if doesFileExist(path.Join(rootPath, "vendor")) {
 		// run ensure
-		cmd := exec.Command("dep", "ensure", "-v")
+		cmd := exec.Command("dep", "ensure", "-v") // nolint: gas
 		cmd.Dir = rootPath
 		if b, err := cmd.CombinedOutput(); err != nil {
 			return fmt.Errorf("error while running dep ensure on %s: %s - %s", rootPath, string(b), err.Error())
 		}
 	} else {
 		// first time, run init
-		cmd := exec.Command("dep", "init", "-v")
+		cmd := exec.Command("dep", "init", "-v") // nolint: gas
 		cmd.Dir = rootPath
 		if b, err := cmd.CombinedOutput(); err != nil {
 			return fmt.Errorf("error while running dep init on %s: %s - %s", rootPath, string(b), err.Error())
 		}
 	}
-
 	return nil
 }
